C++ output now compiles with gcc-x86-4.7.0.

+++ intlibs/idl-common/types.cpp
@@ -209,9 +209,9 @@ string getSJType(const Interface& inf, const string& type, bool argIn) {
 	if(argType == "noreturn")
 		return "void";
 	if(argType == "float")
-		return "int";
+		return "double";
 	if(argType == "double")
-		return "int";
+		return "double";
 
 	if(argType == "uint")
 		return "int";
